Refactor shiki transformers and update import paths
