+++ www/docs/src/components/PackageManagers.astro
@@ -2,7 +2,7 @@
 import { Code } from 'astro:components';
 import { TabItem, Tabs } from '@astrojs/starlight/components';
 import type { ComponentProps } from 'astro/types';
-import { addCopyButton } from '~/shikiTransformers/copyButton';
+import { addCopyButton } from '~/shiki/transformers/copyButton';
 
 import {
 	type CommandOptions,
