+++ www/docs/src/shiki/transformers/colorizedBrackets.ts
