+++ www/docs/src/shiki/transformers/copyButton.ts
@@ -24,11 +24,19 @@ export function addCopyButton(options?: { toggle?: number }): ShikiTransformer {
 					h('span', { class: 'ready' }, [
 						h(
 							'svg',
-							{ xmlns: 'http://www.w3.org/2000/svg', fill: 'none', width: 24, height: 24, viewBox: '0 0 24 24', stroke: 'var(--copy-button)', strokeWidth: "1.5" },
+							{
+								xmlns: 'http://www.w3.org/2000/svg',
+								fill: 'none',
+								width: 24,
+								height: 24,
+								viewBox: '0 0 24 24',
+								stroke: 'var(--copy-button)',
+								strokeWidth: '1.5',
+							},
 							[
 								h('path', {
-									strokeLinecap: "round",
-									strokeLinejoin: "round",
+									strokeLinecap: 'round',
+									strokeLinejoin: 'round',
 									d: 'M15.75 17.25v3.375c0 .621-.504 1.125-1.125 1.125h-9.75a1.125 1.125 0 0 1-1.125-1.125V7.875c0-.621.504-1.125 1.125-1.125H6.75a9.06 9.06 0 0 1 1.5.124m7.5 10.376h3.375c.621 0 1.125-.504 1.125-1.125V11.25c0-4.46-3.243-8.161-7.5-8.876a9.06 9.06 0 0 0-1.5-.124H9.375c-.621 0-1.125.504-1.125 1.125v3.5m7.5 10.375H9.375a1.125 1.125 0 0 1-1.125-1.125v-9.25m12 6.625v-1.875a3.375 3.375 0 0 0-3.375-3.375h-1.5a1.125 1.125 0 0 1-1.125-1.125v-1.5a3.375 3.375 0 0 0-3.375-3.375H9.75',
 								}),
 							]
@@ -37,11 +45,19 @@ export function addCopyButton(options?: { toggle?: number }): ShikiTransformer {
 					h('span', { class: 'success' }, [
 						h(
 							'svg',
-							{ xmlns: 'http://www.w3.org/2000/svg', width: 24, height: 24, viewBox: '0 0 24 24', strokeWidth: "1.5", fill: "none", stroke: 'var(--copy-button-success)' },
+							{
+								xmlns: 'http://www.w3.org/2000/svg',
+								width: 24,
+								height: 24,
+								viewBox: '0 0 24 24',
+								strokeWidth: '1.5',
+								fill: 'none',
+								stroke: 'var(--copy-button-success)',
+							},
 							[
 								h('path', {
-									strokeLinecap: "round",
-									strokeLinejoin: "round",
+									strokeLinecap: 'round',
+									strokeLinejoin: 'round',
 									d: 'm4.5 12.75 6 6 9-13.5',
 								}),
 							]
@@ -54,3 +70,5 @@ export function addCopyButton(options?: { toggle?: number }): ShikiTransformer {
 		},
 	};
 }
+
+export default addCopyButton;
