+++ www/docs/src/shiki/transformers/index.ts
@@ -1,0 +1,3 @@
+export { default as metaTitle } from './metaTitle';
+export { default as colorizedBrackets } from './colorizedBrackets';
+export { default as addCopyButton } from './copyButton';
