+++ www/docs/src/shiki/transformers/metaTitle.ts
@@ -37,3 +37,5 @@ export function metaTitle(): ShikiTransformer {
 		},
 	};
 }
+
+export default metaTitle;
